% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_axis.R
\name{adapt_axis}
\alias{adapt_axis}
\title{A function to define axis in a ggplot of stable istopes}
\usage{
adapt_axis(
  SIA_plot = SIA_plot,
  C_breaks = C_breaks,
  C_limits = C_limits,
  N_breaks = N_breaks,
  N_limits = N_limits
)
}
\arguments{
\item{SIA_plot}{the basic plot}

\item{C_breaks}{the breaks for the axis containing carbon values}

\item{C_limits}{the limits for the axis containing carbon values}

\item{N_breaks}{the breaks for the axis contianing nitrogen values}

\item{N_limits}{the limits for the axis containing nitrogen values}
}
\value{
a ggplot2 with defined axis x and y
}
\description{
A function to define axis in a ggplot of stable istopes
}
\examples{
SIA_plot_axis<-adapt_axis(SIA_plot=SIA_plot,
C_breaks=c(-20,-19,-18,-17,-16,-15,-14),C_limits=c(-20,-14),
N_breaks=c(14,15,16,17,18),N_limits=c(13.5,18.1))
}
