% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_colors.R
\name{adapt_colors}
\alias{adapt_colors}
\title{A function to change the color of the ellipses in ggplot}
\usage{
adapt_colors(SIA_plot = SIA_plot, color1 = color1, color2 = color2)
}
\arguments{
\item{SIA_plot}{a ggplot (will show as a list of 9)}

\item{color1}{the color for group 1}

\item{color2}{the color for group 2}

\item{column_group}{the name of the column group}
}
\value{
a ggplot with ellipses in different colors
}
\description{
A function to change the color of the ellipses in ggplot
}
\examples{
SIA_plot_colors<-adapt_colors(SIA_plot=SIA_plot,
color1='#233d4d',color2='#2a9d8f',column_group='Species')
}
