% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_legend.R
\name{adapt_legend}
\alias{adapt_legend}
\title{A function to add text for labels inside the plot}
\usage{
adapt_legend(
  SIA_plot = SIA_plot,
  xlabel_adjust = xlabel_adjust,
  ylabel_adjust = ylabel_adjust,
  label1 = label1,
  label2 = label2,
  color1 = color1,
  color2 = color2
)
}
\arguments{
\item{SIA_plot}{is a ggplot2 with the SIA ellipses}

\item{xlabel_adjust}{is a number on the x axis}

\item{ylabel_adjust}{is a number on the y axis}

\item{label1}{is the text that will be displayed upper}

\item{label2}{is the text that will be displayed under}

\item{color1}{is the color of the circle displayed upper}

\item{color2}{is the color of the circle displayed under}
}
\value{
a ggplot with text as legend
}
\description{
A function to add text for labels inside the plot
}
\examples{
SIA_plot_legend<-adapt_legend(SIA_plot=SIA_plot,
xlabel_adjust=-19.9,ylabel_adjust=17.9,label1='Masked booby',label2='Red-footed booby',
color1='#233d4d',color2='#2a9d8f')
}
