% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_theme.R
\name{adapt_theme}
\alias{adapt_theme}
\title{A function to add theme for SIA ggplot2}
\usage{
adapt_theme(SIA_plot = SIA_plot)
}
\arguments{
\item{SIA_plot}{SIA plot created in ggplot2 it will show as a list}
}
\value{
a ggplot with a specific theme
}
\description{
A function to add theme for SIA ggplot2
}
\examples{
SIA_plot_theme<-adapt_theme(SIA_plot=SIA_plot)
}
