% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas_overlap.R
\name{areas_overlap}
\alias{areas_overlap}
\title{A function to directly calculate areas of overlap between ellipses using functions from the package SIBER}
\usage{
areas_overlap(
  SIBER_object = SIBER_object,
  group1 = group1,
  group2 = group2,
  overlap_value = overlap_value
)
}
\arguments{
\item{SIBER_object}{A SIBER object}

\item{group1}{the name of the group 1}

\item{group2}{the name of the group 2}

\item{overlap_value}{the overlap you want to calculate, e.g. 0.50 50, 0.95 95}
}
\value{
returns three values, 1. area overlap of group 1, 2. area of overlap of group 2 and 3. area of overlap between the two
}
\description{
A function to directly calculate areas of overlap between ellipses using functions from the package SIBER
}
\examples{
SIBER_overlap_50<-areas_overlap(SIBER_object = SIBER_object,
group1='SulaSula',group2='Dactylatra',overlap_value = 0.50)
}
