% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dispersion.R
\name{calculate_dispersion}
\alias{calculate_dispersion}
\title{A function to calculate dispersion}
\usage{
calculate_dispersion(SIA_residuals = SIA_residuals, SIA_group = SIA_group)
}
\arguments{
\item{SIA_residuals}{is the residuals from carbon and nitrogen}

\item{SIA_group}{is the groups in the model (as factor)}
}
\value{
the dispersion calculations
}
\description{
A function to calculate dispersion
}
\examples{
SIA_dispersion<-calculate_dispersion(SIA_residuals = SIA_residuals,SIA_group=SIA_group)
}
