% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hotellings.R
\name{calculate_hotellings}
\alias{calculate_hotellings}
\title{Calculate hotellings T2 using Turner et al 2020 script
'Hotellings reveleaded that Euclidian distance between centroids differed significantly from zero for contrast between guilds'}
\usage{
calculate_hotellings(
  SIA_means = SIA_means,
  SIA_group = SIA_group,
  SIA_matrix = SIA_matrix
)
}
\arguments{
\item{SIA_means}{the means from your calculations}

\item{SIA_group}{group levels as factors}

\item{SIA_matrix}{a matrix with carbon and nitrogen values}
}
\value{
values hotellings
}
\description{
Calculate hotellings T2 using Turner et al 2020 script
'Hotellings reveleaded that Euclidian distance between centroids differed significantly from zero for contrast between guilds'
}
\examples{
SIA_hotellings<-calculate_hotellings(SIA_means=calculate_means(SIA_matrix,SIA_group),
SIA_group=SIA_group,
SIA_matrix=SIA_matrix)
}
