% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_means.R
\name{calculate_means}
\alias{calculate_means}
\title{A function to calculate group means}
\usage{
calculate_means(SIA_matrix = SIA_matrix, SIA_group = SIA_group)
}
\arguments{
\item{SIA_matrix}{A matrix with carbon and nitrogen values}

\item{SIA_group}{Groups as factors}
}
\value{
Means per group
}
\description{
A function to calculate group means
}
\examples{
SIA_means<-calculate_means(SIA_matrix,SIA_group)
}
