% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvalues.R
\name{calculate_pvalues}
\alias{calculate_pvalues}
\title{Calculate the P-value by measuring the rank of the first value in the distribution of values in the same column.}
\usage{
calculate_pvalues(SIA_permutation = SIA_permutation)
}
\arguments{
\item{SIA_permutation}{results from the permutation producedure}
}
\value{
p values
}
\description{
Calculate the P-value by measuring the rank of the first value in the distribution of values in the same column.
}
\examples{
SIA_pvalue<-calculate_pvalues(SIA_permutation=SIA_permutation)
}
