% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_residuals.R
\name{calculate_residuals}
\alias{calculate_residuals}
\title{A function to calculate the residuals from the models}
\usage{
calculate_residuals(SIA_matrix = SIA_matrix, SIA_group = SIA_group)
}
\arguments{
\item{SIA_matrix}{Contains the carbon and the nitrogen values}

\item{SIA_group}{Contains the group values}
}
\value{
Returns two columns, one for residuals from carbon and other column for nitrogen
}
\description{
A function to calculate the residuals from the models
}
\examples{
SIA_residuals<-calculate_residuals(SIA_matrix=SIA_matrix,SIA_group = SIA_group)
}
