% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_differences.R
\name{contrast_differences}
\alias{contrast_differences}
\title{A function to contrast mean differences}
\usage{
contrast_differences(SIA_tabledispersion = SIA_tabledispersion)
}
\arguments{
\item{SIA_tabledispersion}{A table with mean dispersion values}
}
\value{
A data frame with dispersion differences using different methods
}
\description{
A function to contrast mean differences
}
\examples{
SIA_dispersiondif<-contrast_differences(SIA_tabledispersion = SIA_tabledispersion)
}
