% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_object.R
\name{convert_object}
\alias{convert_object}
\title{A function to convert my data frame into siber object}
\usage{
convert_object(SIBER_SIA = SIBER_SIA)
}
\arguments{
\item{SIBER_SIA}{the organized data frame is not going to be converted to siber object for later uses}
}
\value{
the siber object as requiered by SIBER
}
\description{
A function to convert my data frame into siber object
}
\examples{
SIBER_object<-convert_object(SIBER_SIA=SIBER_SIA)
}
