% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Create matrix to run functions from Turner et al. 2010}
\usage{
create_matrix(SIA_Turner = SIA_Turner)
}
\arguments{
\item{SIA_Turner}{A data frame containing the carbon, nitrogen and group values}
}
\value{
SIA_matrix
}
\description{
Create matrix to run functions from Turner et al. 2010
}
\examples{
SIA_matrix<-create_matrix(SIA_Turner = SIA_Turner)
}
