% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_statistics.R
\name{dispersion_statistics}
\alias{dispersion_statistics}
\title{A function to calculate dispersion statistics}
\usage{
dispersion_statistics(SIA_dispersion = SIA_dispersion, SIA_group = SIA_group)
}
\arguments{
\item{SIA_dispersion}{Are values calculated with the function calculate_dispersion}

\item{SIA_group}{is the group levels as factor}
}
\value{
returns a table with values
}
\description{
A function to calculate dispersion statistics
}
\examples{
SIA_tabledispersion<-dispersion_statistics(SIA_dispersion=SIA_dispersion,SIA_group=SIA_group)
}
