% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ellipse.R
\name{ggplot_ellipse}
\alias{ggplot_ellipse}
\title{A function to plot ellipses using ggplot2}
\usage{
ggplot_ellipse(
  SIA_samples = SIA_samples,
  column_C = column_C,
  column_N = column_N,
  column_group = column_group
)
}
\arguments{
\item{SIA_samples}{A data frame with a column for carbon, nitrogen and group}

\item{column_C}{the name of the column carbon}

\item{column_N}{the name of the column nitrogen}

\item{column_group}{the name of the column containing the groups}
}
\value{
returns a ggplot2 that can be further modified, plots 50 and 95 ellipses
}
\description{
A function to plot ellipses using ggplot2
}
\examples{
SIA_plot<-ggplot_ellipse(SIA_samples<-Sula_data, column_C<-'Carbon',
column_N<-'Nitrogeno', column_group<-'Species')
}
