% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_for_SIBER.R
\name{organize_for_SIBER}
\alias{organize_for_SIBER}
\title{A function to organize columns as requiered by SIBER}
\usage{
organize_for_SIBER(
  SIA_samples = SIA_samples,
  column_N = column_N,
  column_C = column_C,
  column_group = column_group
)
}
\arguments{
\item{SIA_samples}{A data frame with the carbon, nitrogen and group values}

\item{column_N}{the name of the column that contains the nitrogen values}

\item{column_C}{the name of the column that contains the carbon values}

\item{column_group}{the name of the column that contains the groups}
}
\value{
a new data frame organized as requiered by SIBER
}
\description{
A function to organize columns as requiered by SIBER
}
\examples{
SIBER_SIA<-organize_for_SIBER(SIA_samples=Sula_data,
column_C='Carbon',column_N='Nitrogeno',column_group='Species')
}
