% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_for_Turner.R
\name{organize_for_Turner}
\alias{organize_for_Turner}
\title{reorganize_for_Turner}
\usage{
organize_for_Turner(
  SIA_samples = SIA_samples,
  column_C = column_C,
  column_N = column_N,
  column_group = column_group
)
}
\arguments{
\item{SIA_samples}{data frame with carbon, nitrogen and group}

\item{column_C}{the name of the column that contains the carbon values}

\item{column_N}{the name of the column that contains the nitrogen values}

\item{column_group}{the name of the column that contains the group}
}
\value{
your data frame organized to be used in posterior functions
}
\description{
reorganize_for_Turner
}
\examples{
SIA_Turner<-organize_for_Turner(SIA_samples=Sula_data,
column_C='Carbon',column_N='Nitrogeno',column_group='Species')
}
