% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_centroids.R
\name{plot_centroids}
\alias{plot_centroids}
\title{A function to plot the stable isotopes values and the centroid locations}
\usage{
plot_centroids(
  SIA_samples = SIA_samples,
  column_C = column_C,
  column_N = column_N,
  column_group = column_group,
  group1 = group1,
  group2 = group2
)
}
\arguments{
\item{SIA_samples}{a data frame with your stable isotope data}

\item{column_C}{the name of the column that contains the carbon values}

\item{column_N}{the name of the column that contains the nitrogen values}

\item{column_group}{the name of the column with the groups to compare}

\item{group1}{the name (in characters) of the first group}

\item{group2}{the name (in characters) of the second group}
}
\value{
prints a plot in base R
}
\description{
A function to plot the stable isotopes values and the centroid locations
}
\examples{
plot_centroids(SIA_samples=Sula_data, column_C='Carbon', column_N='Nitrogeno',
column_group='Species', group1="SulaSula", group2="Dactylatra")
}
