% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_overlap.R
\name{proportion_overlap}
\alias{proportion_overlap}
\title{A function to calculate the proportion of overlap between areas}
\usage{
proportion_overlap(area_overlap = area_overlap)
}
\arguments{
\item{area_overlap}{is the overlap that you want to compare, it needs to have three columns}
}
\value{
the proportion of overlap in a two digits decimals
}
\description{
A function to calculate the proportion of overlap between areas
}
\examples{
prop_overlap_95<-proportion_overlap(area_overlap = SIBER_overlap_95)
}
