% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_permutation.R
\name{run_permutation}
\alias{run_permutation}
\title{A function for a permutation procedure}
\usage{
run_permutation(
  SIA_matrix = SIA_matrix,
  SIA_group = SIA_group,
  SIA_dispersiondif = SIA_dispersiondif,
  permutation_runs = permutation_runs
)
}
\arguments{
\item{SIA_matrix}{A matrix with carbon and nitrogen values}

\item{SIA_group}{Group levels as factors}

\item{SIA_dispersiondif}{Calculation from dispersions}

\item{permutation_runs}{Number of permutations}
}
\value{
results from permutation
}
\description{
A function for a permutation procedure
}
\examples{
SIA_permutation<-run_permutation(SIA_matrix=SIA_matrix, SIA_group=SIA_group,
SIA_dispersiondif = SIA_dispersiondif,permutation_runs=999)
}
