% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_dispersion.R
\docType{data}
\name{SIA_dispersion}
\alias{SIA_dispersion}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Values of dispersion}
\format{
Numeric values
}
\usage{
data(SIA_dispersion)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Values of dispersion
}
\examples{
data(SIA_dispersion)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
