% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_dispersiondif.R
\docType{data}
\name{SIA_dispersiondif}
\alias{SIA_dispersiondif}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
List of 3: mdc, mnn, ecc. Calculated using Turners et al. 2010 workflow}
\format{
List with 3 elements
}
\usage{
data(SIA_dispersiondif)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
List of 3: mdc, mnn, ecc. Calculated using Turners et al. 2010 workflow
}
\examples{
data(SIA_dispersiondif)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
