% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_group.R
\docType{data}
\name{SIA_group}
\alias{SIA_group}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Factor with two levels. As requiered for using Turner et al. 2010 workflow}
\format{
Factos with two levels
}
\usage{
data(SIA_group)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Factor with two levels. As requiered for using Turner et al. 2010 workflow
}
\examples{
data(SIA_group)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
