% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_matrix.R
\docType{data}
\name{SIA_matrix}
\alias{SIA_matrix}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Matrix with two columns and 59 observations. As requiered for using Turner et al. 2010 workflow}
\format{
Matrix with two columns and 59 observations
}
\usage{
data(SIA_matrix)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Matrix with two columns and 59 observations. As requiered for using Turner et al. 2010 workflow
}
\examples{
data(SIA_matrix)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
