% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_permutation.R
\docType{data}
\name{SIA_permutation}
\alias{SIA_permutation}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Results from iteration. As requiered for using Turner et al. 2010 workflow}
\format{
Iterations of 1 to 5005
}
\usage{
data(SIA_permutation)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Results from iteration. As requiered for using Turner et al. 2010 workflow
}
\examples{
data(SIA_permutation)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
