% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_plot.R
\docType{data}
\name{SIA_plot}
\alias{SIA_plot}
\title{Ellipses plot from data of red-footed and masked boobies collected at Isla Clarion, Mexico
Create ellipses in ggplot2.}
\format{
List of 9 (ggplot format)
}
\usage{
data(SIA_plot)
}
\description{
Ellipses plot from data of red-footed and masked boobies collected at Isla Clarion, Mexico
Create ellipses in ggplot2.
}
\examples{
data(SIA_plot)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
