% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_residuals.R
\docType{data}
\name{SIA_residuals}
\alias{SIA_residuals}
\title{Residuals form data of red-footed and masked boobies collected at Isla Clarion, Mexico
Calculated residuals using functions from Turner et al 2010.}
\format{
List of 9 (ggplot format)
}
\usage{
data(SIA_residuals)
}
\description{
Residuals form data of red-footed and masked boobies collected at Isla Clarion, Mexico
Calculated residuals using functions from Turner et al 2010.
}
\examples{
data(SIA_residuals)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
