% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_tabledispersion.R
\docType{data}
\name{SIA_tabledispersion}
\alias{SIA_tabledispersion}
\title{Dispersion measurements form data of red-footed and masked boobies collected at Isla Clarion, Mexico
Tabla created using functions from Turner et al 2010.}
\format{
Object with two rows and three columns
}
\usage{
data(SIA_tabledispersion)
}
\description{
Dispersion measurements form data of red-footed and masked boobies collected at Isla Clarion, Mexico
Tabla created using functions from Turner et al 2010.
}
\examples{
data(SIA_tabledispersion)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
