% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIA_Turner.R
\docType{data}
\name{SIA_Turner}
\alias{SIA_Turner}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Contains 59 observations and 3 variables. Its organized as requiered for Turners et al. 2010 workflow}
\format{
Data frame with 3 variables
}
\usage{
data(SIA_Turner)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Contains 59 observations and 3 variables. Its organized as requiered for Turners et al. 2010 workflow
}
\examples{
data(SIA_Turner)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
