% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIBER_object.R
\docType{data}
\name{SIBER_object}
\alias{SIBER_object}
\title{List of 12}
\format{
object produced by SIBER package
}
\usage{
data(SIBER_object)
}
\description{
List of 12
}
\examples{
data(SIBER__object)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
