% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIBER_overlap_95.R
\docType{data}
\name{SIBER_overlap_95}
\alias{SIBER_overlap_95}
\title{Data from overlaps of 95 returns numerical values
Contains numerical values from the overlaps, one from area 1 , one from area 2 and one from total area}
\format{
Numerical values
}
\usage{
data(SIBER_overlap_95)
}
\description{
Data from overlaps of 95 returns numerical values
Contains numerical values from the overlaps, one from area 1 , one from area 2 and one from total area
}
\examples{
data(SIBER_overlap_95)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
