% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIBER_SIA.R
\docType{data}
\name{SIBER_SIA}
\alias{SIBER_SIA}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Contains 59 observations and 3 variables. Its organized as requiered for SIBER workflow}
\format{
Data frame with 3 variables
}
\usage{
data(SIBER_SIA)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Contains 59 observations and 3 variables. Its organized as requiered for SIBER workflow
}
\examples{
data(SIBER_SIA)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
