% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sula_data.R
\docType{data}
\name{Sula_data}
\alias{Sula_data}
\title{Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Contains 59 observations and 3 variables.}
\format{
Data frame with 3 variables
}
\usage{
data(Sula_data)
}
\description{
Data from red-footed and masked boobies collected at Isla Clarion, Mexico
Contains 59 observations and 3 variables.
}
\examples{
data(Sula_data)
}
\references{
Lerma et al. 2023
}
\keyword{datasets}
